src/backend/requirements.txt
appscript==1.2.5
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
numpy==2.0.0
pandas==2.2.2
pdfrw==0.4
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
